heuristics abstract class created, HC and RS working the Problem class and  removed some unused imports
