+++ examples/santafe.py
@@ -192,10 +192,15 @@ def evolve(
         representation = dsge_representation
     else:
         representation = treebased_representation
+
     alg = GP(
         g,
-        fitness_function,
         representation=representation,
+        problem=SingleObjectiveProblem(
+            minimize=False,
+            fitness_function=fitness_function,
+            target_fitness=None,
+        ),
         # number_of_generations=150,
         # population_size=100,
         # max_depth=15,
@@ -203,7 +208,6 @@ def evolve(
         # probability_crossover=0.75,
         # probability_mutation=0.01,
         # selection_method=("tournament", 2),
-        minimize=False,
         seed=seed,
         timer_stop_criteria=mode,
     )
