+++ geneticengine/core/problems.py
@@ -2,9 +2,6 @@ from __future__ import annotations
 
 from abc import ABC
 from dataclasses import dataclass
-from pickle import bytes_types
-from tokenize import Single
-from typing import Any
 from typing import Callable
 from typing import Optional
 from typing import TypeVar
